+++ close-milestone/index.ts
@@ -1,19 +1,13 @@
 import { context } from '@actions/github'
 import { Action } from '../common/Action'
 import { OctoKit } from '../api/octokit'
-import { EventPayloads } from '@octokit/webhooks'
 
 class CloseMilestone extends Action {
 	id = 'CloseMilestone'
 
 	async onTriggered(octokit: OctoKit) {
 		const { owner, repo } = context.repo
-		const payload = context.payload as EventPayloads.WebhookPayloadWorkflowDispatch
-		const version = (payload.inputs as any).version
-
-		if (!version) {
-			throw new Error('Missing version input')
-		}
+		const version = this.getVersion()
 
 		// get all the milestones
 		const milestones = await octokit.octokit.issues.listMilestonesForRepo({
